migrate notify code to separate module
