+++ src/maestral/notify/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from .notify import MaestralDesktopNotifier, MaestralDesktopNotificationHandler
+
+__all__ = ["MaestralDesktopNotifier", "MaestralDesktopNotificationHandler"]
