+++ src/maestral/notify/notify_base.py
